Style cleanup, gitignore fix
